move plugin to module
