--- bfe_plugin/plugin_info.go
@@ -1,29 +1,0 @@
-// Copyright (c) 2019 Baidu, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package bfe_plugin
-
-import (
-	"github.com/baidu/bfe/bfe_module"
-	"github.com/baidu/go-lib/web-monitor/web_monitor"
-)
-
-type PluginInfo struct {
-	Name        string
-	Version     string
-	Description string
-	Kind        string
-	Path        string
-	Init        func(cbs *bfe_module.BfeCallbacks, whs *web_monitor.WebHandlers, cr string) error
-}
